MAINT: refactor Fourier module into smaller parts
