+++ odl/test/trafos/fourier_test.py
@@ -22,17 +22,17 @@ from __future__ import print_function, division, absolute_import
 from future import standard_library
 standard_library.install_aliases()
 
-# External module imports
 from itertools import product
 from math import pi
 import numpy as np
 import pytest
 
-# ODL imports
 import odl
-from odl.trafos.fourier import (
+from odl.trafos.backends import pyfftw_call
+from odl.trafos.util.ft_utils import (
     reciprocal, inverse_reciprocal, dft_preprocess_data, dft_postprocess_data,
-    pyfftw_call, _interp_kernel_ft,
+    _interp_kernel_ft)
+from odl.trafos.fourier import (
     DiscreteFourierTransform, DiscreteFourierTransformInverse,
     FourierTransform)
 from odl.util import (all_almost_equal, all_equal,
