+++ odl/trafos/backends/__init__.py
@@ -15,21 +15,12 @@
 # You should have received a copy of the GNU General Public License
 # along with ODL.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Utility functions for transformations."""
+"""Bindings to external backends for transformations."""
 
 from __future__ import absolute_import
 
 
 __all__ = ()
 
-from . import util
-
-from . import backends
-from .backends.pyfftw_bindings import PYFFTW_AVAILABLE
-__all__ += (PYFFTW_AVAILABLE,)
-
-from .fourier import *
-__all__ += fourier.__all__
-
-from .wavelet import *
-__all__ += wavelet.__all__
+from . pyfftw_bindings import *
+__all__ += pyfftw_bindings.__all__
