+++ odl/trafos/fourier.py
@@ -23,506 +23,23 @@ from future import standard_library
 standard_library.install_aliases()
 from builtins import range, super
 range_seq = range  # avoid duplicate names
-from future.utils import raise_from
 
-from multiprocessing import cpu_count
 import numpy as np
-try:
-    import pyfftw
-    PYFFTW_AVAILABLE = True
-except ImportError:
-    PYFFTW_AVAILABLE = False
 
-from odl.discr import (RegularGrid, DiscreteLp,
-                       discr_sequence_space, uniform_partition_fromgrid,
-                       uniform_discr_frompartition)
+from odl.discr import RegularGrid, DiscreteLp, discr_sequence_space
 from odl.operator import Operator
 from odl.set import RealNumbers, ComplexNumbers
-from odl.util import (fast_1d_tensor_mult, is_real_dtype, is_scalar_dtype,
-                      is_real_floating_dtype, is_complex_floating_dtype,
+from odl.trafos.backends.pyfftw_bindings import (
+    pyfftw_call, PYFFTW_AVAILABLE, _pyfftw_to_local)
+from odl.trafos.util import (
+    reciprocal, reciprocal_space, dft_preprocess_data, dft_postprocess_data)
+from odl.util import (is_real_dtype, is_complex_floating_dtype,
                       dtype_repr, conj_exponent, TYPE_MAP_R2C,
                       normalized_scalar_param_list)
 
 
-__all__ = ('FourierTransform', 'FourierTransformInverse',
-           'DiscreteFourierTransform', 'DiscreteFourierTransformInverse',
-           'PYFFTW_AVAILABLE')
-
-
-def reciprocal(grid, shift=True, axes=None, halfcomplex=False):
-    """Return the reciprocal of the given regular grid.
-
-    This function calculates the reciprocal (Fourier/frequency space)
-    grid for a given regular grid defined by the nodes::
-
-        x[k] = x[0] + k * s,
-
-    where ``k = (k[0], ..., k[d-1])`` is a ``d``-dimensional index in
-    the range ``0 <= k < N`` (component-wise). The multi-index
-    ``N`` is the shape of the input grid.
-    This grid's reciprocal is then given by the nodes::
-
-        xi[j] = xi[0] + j * sigma,
-
-    with the reciprocal grid stride ``sigma = 2*pi / (s * N)``.
-    The minimum frequency ``xi[0]`` can in principle be chosen
-    freely, but usually it is chosen in a such a way that the reciprocal
-    grid is centered around zero. For this, there are two possibilities:
-
-    1. Make the grid point-symmetric around 0.
-
-    2. Make the grid "almost" point-symmetric around zero by shifting
-       it to the left by half a reciprocal stride.
-
-    In the first case, the minimum frequency (per axis) is given as::
-
-        xi_1[0] = -pi/s + pi/(s*n) = -pi/s + sigma/2.
-
-    For the second case, it is::
-
-        xi_1[0] = -pi / s.
-
-    Note that the zero frequency is contained in case 1 for an odd
-    number of points, while for an even size, the second option
-    guarantees that 0 is contained.
-
-    If a real-to-complex (half-complex) transform is to be computed,
-    the reciprocal grid has the shape ``M[i] = floor(N[i]/2) + 1``
-    in the last transform axis ``i``.
-
-    Parameters
-    ----------
-    grid : `RegularGrid`
-        Original sampling grid
-    shift : bool or `sequence` of bools, optional
-        If ``True``, the grid is shifted by half a stride in the negative
-        direction. With a sequence, this option is applied separately on
-        each axis.
-    axes : int or `sequence` of ints, optional
-        Dimensions in which to calculate the reciprocal. The sequence
-        must have the same length as ``shift`` if the latter is given
-        as a sequence. ``None`` means all axes in ``grid``.
-    halfcomplex : bool, optional
-        If ``True``, return the half of the grid with last coordinate
-        less than zero. This is related to the fact that for real-valued
-        functions, the other half is the mirrored complex conjugate of
-        the given half and therefore needs not be stored.
-
-    Returns
-    -------
-    recip : `RegularGrid`
-        The reciprocal grid
-    """
-    if axes is None:
-        axes = list(range(grid.ndim))
-    else:
-        try:
-            axes = [int(axes)]
-        except TypeError:
-            axes = list(axes)
-
-    # List indicating shift or not per "active" axis, same length as axes
-    shift_list = normalized_scalar_param_list(shift, length=len(axes),
-                                              param_conv=bool)
-
-    # Full-length vectors
-    stride = grid.stride
-    shape = np.array(grid.shape)
-    rmin = grid.min_pt.copy()
-    rmax = grid.max_pt.copy()
-    rshape = list(shape)
-
-    # Shifted axes (full length to avoid ugly double indexing)
-    shifted = np.zeros(grid.ndim, dtype=bool)
-    shifted[axes] = shift_list
-    rmin[shifted] = -np.pi / stride[shifted]
-    # Length min->max increases by double the shift, so we
-    # have to compensate by a full stride
-    rmax[shifted] = (-rmin[shifted] -
-                     2 * np.pi / (stride[shifted] * shape[shifted]))
-
-    # Non-shifted axes
-    not_shifted = np.zeros(grid.ndim, dtype=bool)
-    not_shifted[axes] = np.logical_not(shift_list)
-    rmin[not_shifted] = ((-1.0 + 1.0 / shape[not_shifted]) *
-                         np.pi / stride[not_shifted])
-    rmax[not_shifted] = -rmin[not_shifted]
-
-    # Change last axis shape and max if halfcomplex
-    if halfcomplex:
-        rshape[axes[-1]] = shape[axes[-1]] // 2 + 1
-
-        # - Odd and not shifted or even and shifted -> 0
-        # - Odd and shifted -> - stride / 2
-        # - Even and not shifted -> + stride / 2
-        last_odd = shape[axes[-1]] % 2 == 1
-        last_shifted = shift_list[-1]
-        half_rstride = np.pi / (shape[axes[-1]] * stride[axes[-1]])
-
-        if last_odd:
-            if last_shifted:
-                rmax[axes[-1]] = -half_rstride
-            else:
-                rmax[axes[-1]] = 0
-        else:
-            if last_shifted:
-                rmax[axes[-1]] = 0
-            else:
-                rmax[axes[-1]] = half_rstride
-
-    return RegularGrid(rmin, rmax, rshape)
-
-
-def inverse_reciprocal(grid, x0, axes=None, halfcomplex=False,
-                       halfcx_parity='even'):
-    """Return the inverse reciprocal of the given regular grid.
-
-    Given a reciprocal grid::
-
-        xi[j] = xi[0] + j * sigma,
-
-    with a multi-index ``j = (j[0], ..., j[d-1])`` in the range
-    ``0 <= j < M``, this function calculates the original grid::
-
-        x[k] = x[0] + k * s
-
-    by using a provided ``x[0]`` and calculating the stride ``s``.
-
-    If the reciprocal grid is interpreted as coming from a usual
-    complex-to-complex FFT, it is ``N == M``, and the stride is::
-
-        s = 2*pi / (sigma * N)
-
-    For a reciprocal grid from a real-to-complex (half-complex) FFT,
-    it is ``M[i] = floor(N[i]/2) + 1`` in the last transform axis ``i``.
-    To resolve the ambiguity regarding the parity of ``N[i]``, the
-    it must be specified if the output shape should be even or odd,
-    resulting in::
-
-        odd : N[i] = 2 * M[i] - 1
-        even: N[i] = 2 * M[i] - 2
-
-    The output stride is calculated with this ``N`` as above in this
-    case.
-
-    Parameters
-    ----------
-    grid : `RegularGrid`
-        Original sampling grid
-    x0 : `array-like`
-        Minimal point of the inverse reciprocal grid
-    axes : int or `sequence` of ints, optional
-        Dimensions in which to calculate the reciprocal. The sequence
-        must have the same length as ``shift`` if the latter is given
-        as a sequence. ``None`` means all axes in ``grid``.
-    halfcomplex : bool, optional
-        If ``True``, interpret the given grid as the reciprocal as used
-        in a half-complex FFT (see above). Otherwise, the grid is
-        regarded as being used in a complex-to-complex transform.
-    halfcx_parity : {'even', 'odd'}
-        Use this parity for the shape of the returned grid in the
-        last axis of ``axes`` in the case ``halfcomplex=True``
-
-    Returns
-    -------
-    irecip : `RegularGrid`
-        The inverse reciprocal grid
-    """
-    if axes is None:
-        axes = list(range(grid.ndim))
-    else:
-        try:
-            axes = [int(axes)]
-        except TypeError:
-            axes = list(axes)
-
-    rstride = grid.stride
-    rshape = grid.shape
-
-    # Calculate shape of the output grid by adjusting in axes[-1]
-    irshape = list(rshape)
-    if halfcomplex:
-        if str(halfcx_parity).lower() == 'even':
-            irshape[axes[-1]] = 2 * rshape[axes[-1]] - 2
-        elif str(halfcx_parity).lower() == 'odd':
-            irshape[axes[-1]] = 2 * rshape[axes[-1]] - 1
-        else:
-            raise ValueError("`halfcomplex` parity '{}' not understood"
-                             "".format(halfcx_parity))
-
-    irmin = np.asarray(x0)
-    irshape = np.asarray(irshape)
-    irstride = np.copy(rstride)
-    irstride[axes] = 2 * np.pi / (irshape[axes] * rstride[axes])
-    irmax = irmin + (irshape - 1) * irstride
-
-    return RegularGrid(irmin, irmax, irshape)
-
-
-def _pyfftw_to_local(flag):
-    return flag.lstrip('FFTW_').lower()
-
-
-def _local_to_pyfftw(flag):
-    return 'FFTW_' + flag.upper()
-
-
-def _pyfftw_destroys_input(flags, direction, halfcomplex, ndim):
-    """Return ``True`` if FFTW destroys an input array, ``False`` otherwise."""
-    if any(flag in flags or _pyfftw_to_local(flag) in flags
-           for flag in ('FFTW_MEASURE', 'FFTW_PATIENT', 'FFTW_EXHAUSTIVE',
-                        'FFTW_DESTROY_INPUT')):
-        return True
-    elif (direction in ('backward', 'FFTW_BACKWARD') and halfcomplex and
-          ndim != 1):
-        return True
-    else:
-        return False
-
-
-def _pyfftw_check_args(arr_in, arr_out, axes, halfcomplex, direction):
-    """Raise an error if anything is not ok with in and out."""
-    if len(set(axes)) != len(axes):
-        raise ValueError('duplicate axes are not allowed')
-
-    if direction == 'forward':
-        out_shape = list(arr_in.shape)
-        if halfcomplex:
-            try:
-                out_shape[axes[-1]] = arr_in.shape[axes[-1]] // 2 + 1
-            except IndexError as err:
-                raise_from(IndexError('axis index {} out of range for array '
-                                      'with {} axes'
-                                      ''.format(axes[-1], arr_in.ndim)),
-                           err)
-
-        if arr_out.shape != tuple(out_shape):
-            raise ValueError('expected output shape {}, got {}'
-                             ''.format(tuple(out_shape), arr_out.shape))
-
-        if is_real_dtype(arr_in.dtype):
-            out_dtype = TYPE_MAP_R2C[arr_in.dtype]
-        elif halfcomplex:
-            raise ValueError('cannot combine halfcomplex forward transform '
-                             'with complex input')
-        else:
-            out_dtype = arr_in.dtype
-
-        if arr_out.dtype != out_dtype:
-            raise ValueError('expected output dtype {}, got {}'
-                             ''.format(dtype_repr(out_dtype),
-                                       dtype_repr(arr_out.dtype)))
-
-    elif direction == 'backward':
-        in_shape = list(arr_out.shape)
-        if halfcomplex:
-            try:
-                in_shape[axes[-1]] = arr_out.shape[axes[-1]] // 2 + 1
-            except IndexError as err:
-                raise_from(IndexError('axis index {} out of range for array '
-                                      'with {} axes'
-                                      ''.format(axes[-1], arr_out.ndim)),
-                           err)
-
-        if arr_in.shape != tuple(in_shape):
-            raise ValueError('expected input shape {}, got {}'
-                             ''.format(tuple(in_shape), arr_in.shape))
-
-        if is_real_dtype(arr_out.dtype):
-            in_dtype = TYPE_MAP_R2C[arr_out.dtype]
-        elif halfcomplex:
-            raise ValueError('cannot combine halfcomplex backward transform '
-                             'with complex output')
-        else:
-            in_dtype = arr_out.dtype
-
-        if arr_in.dtype != in_dtype:
-            raise ValueError('expected input dtype {}, got {}'
-                             ''.format(dtype_repr(in_dtype),
-                                       dtype_repr(arr_in.dtype)))
-
-    else:  # Shouldn't happen
-        raise RuntimeError
-
-
-def pyfftw_call(array_in, array_out, direction='forward', axes=None,
-                halfcomplex=False, **kwargs):
-    """Calculate the DFT with pyfftw.
-
-    The discrete Fourier (forward) transform calcuates the sum::
-
-        f_hat[k] = sum_j( f[j] * exp(-2*pi*1j * j*k/N) )
-
-    where the summation is taken over all indices
-    ``j = (j[0], ..., j[d-1])`` in the range ``0 <= j < N``
-    (component-wise), with ``N`` being the shape of the input array.
-
-    The output indices ``k`` lie in the same range, except
-    for half-complex transforms, where the last axis ``i`` in ``axes``
-    is shortened to ``0 <= k[i] < floor(N[i]/2) + 1``.
-
-    In the backward transform, sign of the the exponential argument
-    is flipped.
-
-    Parameters
-    ----------
-    array_in : `numpy.ndarray`
-        Array to be transformed
-    array_out : `numpy.ndarray`
-        Output array storing the transformed values, may be aligned
-        with ``array_in``.
-    direction : {'forward', 'backward'}
-        Direction of the transform
-    axes : int or `sequence` of ints, optional
-        Dimensions along which to take the transform. ``None`` means
-        using all axis and is equivalent to ``np.arange(ndim)``.
-    halfcomplex : bool, optional
-        If ``True``, calculate only the negative frequency part along the
-        last axis. If ``False``, calculate the full complex FFT.
-        This option can only be used with real input data.
-
-    Other Parameters
-    ----------------
-    fftw_plan : ``pyfftw.FFTW``, optional
-        Use this plan instead of calculating a new one. If specified,
-        the options ``planning_effort``, ``planning_timelimit`` and
-        ``threads`` have no effect.
-    planning_effort : {'estimate', 'measure', 'patient', 'exhaustive'}
-        Flag for the amount of effort put into finding an optimal
-        FFTW plan. See the `FFTW doc on planner flags
-        <http://www.fftw.org/fftw3_doc/Planner-Flags.html>`_.
-        Default: 'estimate'
-    planning_timelimit : float or ``None``, optional
-        Limit planning time to roughly this many seconds.
-        Default: ``None`` (no limit)
-    threads : int, optional
-        Number of threads to use.
-        Default: Number of CPUs if the number of data points is larger
-        than 1000, else 1.
-    normalise_idft : bool, optional
-        If ``True``, the backward transform is normalized by
-        ``1 / N``, where ``N`` is the total number of points in
-        ``array_in[axes]``. This ensures that the IDFT is the true
-        inverse of the forward DFT.
-        Default: ``False``
-    import_wisdom : filename or file handle, optional
-        File to load FFTW wisdom from. If the file does not exist,
-        it is ignored.
-    export_wisdom : filename or file handle, optional
-        File to append the accumulated FFTW wisdom to
-
-    Returns
-    -------
-    fftw_plan : ``pyfftw.FFTW``
-        The plan object created from the input arguments. It can be
-        reused for transforms of the same size with the same data types.
-        Note that reuse only gives a speedup if the initial plan
-        used a planner flag other than ``'estimate'``.
-        If ``fftw_plan`` was specified, the returned object is a
-        reference to it.
-
-    Notes
-    -----
-    * The planning and direction flags can also be specified as
-      capitalized and prepended by ``'FFTW_'``, i.e. in the original
-      FFTW form.
-    * For a ``halfcomplex`` forward transform, the arrays must fulfill
-      ``array_out.shape[axes[-1]] == array_in.shape[axes[-1]] // 2 + 1``,
-      and vice versa for backward transforms.
-    * All planning schemes except ``'estimate'`` require an internal copy
-      of the input array but are often several times faster after the
-      first call (measuring results are cached). Typically,
-      'measure' is a good compromise. If you cannot afford the copy,
-      use ``'estimate'``.
-    * If a plan is provided via the ``fftw_plan`` parameter, no copy
-      is needed internally.
-    """
-    import pickle
-
-    if not array_in.flags.aligned:
-        raise ValueError('input array not aligned')
-
-    if not array_out.flags.aligned:
-        raise ValueError('output array not aligned')
-
-    if axes is None:
-        axes = tuple(range(array_in.ndim))
-    else:
-        try:
-            axes = tuple(int(axes))
-        except TypeError:
-            axes = tuple(axes)
-
-    direction = _pyfftw_to_local(direction)
-    fftw_plan_in = kwargs.pop('fftw_plan', None)
-    planning_effort = _pyfftw_to_local(kwargs.pop('planning_effort',
-                                                  'estimate'))
-    planning_timelimit = kwargs.pop('planning_timelimit', None)
-    threads = kwargs.pop('threads', None)
-    normalise_idft = kwargs.pop('normalise_idft', False)
-    wimport = kwargs.pop('import_wisdom', '')
-    wexport = kwargs.pop('export_wisdom', '')
-
-    # Cast input to complex if necessary
-    array_in_copied = False
-    if is_real_dtype(array_in.dtype) and not halfcomplex:
-        # Need to cast array_in to complex dtype
-        array_in = array_in.astype(TYPE_MAP_R2C[array_in.dtype])
-        array_in_copied = True
-
-    # Do consistency checks on the arguments
-    _pyfftw_check_args(array_in, array_out, axes, halfcomplex, direction)
-
-    # Import wisdom if possible
-    if wimport:
-        try:
-            with open(wimport, 'rb') as wfile:
-                wisdom = pickle.load(wfile)
-        except IOError:
-            wisdom = []
-        except TypeError:  # Got file handle
-            wisdom = pickle.load(wimport)
-
-        if wisdom:
-            pyfftw.import_wisdom(wisdom)
-
-    # Copy input array if it hasn't been done yet and the planner is likely
-    # to destroy it. If we already have a plan, we don't have to worry.
-    planner_destroys = _pyfftw_destroys_input(
-        [planning_effort], direction, halfcomplex, array_in.ndim)
-    must_copy_array_in = fftw_plan_in is None and planner_destroys
-
-    if must_copy_array_in and not array_in_copied:
-        plan_arr_in = np.empty_like(array_in)
-        flags = [_local_to_pyfftw(planning_effort), 'FFTW_DESTROY_INPUT']
-    else:
-        plan_arr_in = array_in
-        flags = [_local_to_pyfftw(planning_effort)]
-
-    if fftw_plan_in is None:
-        if threads is None:
-            if plan_arr_in.size < 1000:  # Somewhat arbitrary
-                threads = 1
-            else:
-                threads = cpu_count()
-
-        fftw_plan = pyfftw.FFTW(
-            plan_arr_in, array_out, direction=_local_to_pyfftw(direction),
-            flags=flags, planning_timelimit=planning_timelimit,
-            threads=threads, axes=axes)
-    else:
-        fftw_plan = fftw_plan_in
-
-    fftw_plan(array_in, array_out, normalise_idft=normalise_idft)
-
-    if wexport:
-        try:
-            with open(wexport, 'ab') as wfile:
-                pickle.dump(pyfftw.export_wisdom(), wfile)
-        except TypeError:  # Got file handle
-            pickle.dump(pyfftw.export_wisdom(), wexport)
-
-    return fftw_plan
+__all__ = ('DiscreteFourierTransform', 'DiscreteFourierTransformInverse',
+           'FourierTransform', 'FourierTransformInverse')
 
 
 class DiscreteFourierTransformBase(Operator):
@@ -1166,404 +683,6 @@ class DiscreteFourierTransformInverse(DiscreteFourierTransformBase):
         return DiscreteFourierTransform(
             domain=self.range, range=self.domain, axes=self.axes,
             halfcomplex=self.halfcomplex, sign=sign)
-
-
-def dft_preprocess_data(arr, shift=True, axes=None, sign='-', out=None):
-    """Pre-process the real-space data before DFT.
-
-    This function multiplies the given data with the separable
-    function::
-
-        p(x) = exp(+- 1j * dot(x - x[0], xi[0]))
-
-    where ``x[0]`` and ``xi[0]`` are the minimum coodinates of
-    the real-space and reciprocal grids, respectively. The sign of
-    the exponent depends on the choice of ``sign``. In discretized
-    form, this function becomes an array::
-
-        p[k] = exp(+- 1j * k * s * xi[0])
-
-    If the reciprocal grid is not shifted, i.e. symmetric around 0,
-    it is ``xi[0] =  pi/s * (-1 + 1/N)``, hence::
-
-        p[k] = exp(-+ 1j * pi * k * (1 - 1/N))
-
-    For a shifted grid, we have :math:``xi[0] =  -pi/s``, thus the
-    array is given by::
-
-        p[k] = (-1)**k
-
-    Parameters
-    ----------
-    arr : `array-like`
-        Array to be pre-processed. If its data type is a real
-        non-floating type, it is converted to 'float64'.
-    shift : bool or or `sequence` of bools, optional
-        If ``True``, the grid is shifted by half a stride in the negative
-        direction. With a sequence, this option is applied separately on
-        each axis.
-    axes : int or `sequence` of ints, optional
-        Dimensions in which to calculate the reciprocal. The sequence
-        must have the same length as ``shift`` if the latter is given
-        as a sequence.
-        Default: all axes.
-    sign : {'-', '+'}, optional
-        Sign of the complex exponent.
-    out : `numpy.ndarray`, optional
-        Array in which the result is stored. If ``out is arr``,
-        an in-place modification is performed. For real data type,
-        this is only possible for ``shift=True`` since the factors are
-        complex otherwise.
-
-    Returns
-    -------
-    out : `numpy.ndarray`
-        Result of the pre-processing. If ``out`` was given, the returned
-        object is a reference to it.
-
-    Notes
-    -----
-    If ``out`` is not specified, the data type of the returned array
-    is the same as that of ``arr`` except when ``arr`` has real data
-    type and ``shift`` is not ``True``. In this case, the return type
-    is the complex counterpart of ``arr.dtype``.
-    """
-    arr = np.asarray(arr)
-    if not is_scalar_dtype(arr.dtype):
-        raise ValueError('array has non-scalar data type {}'
-                         ''.format(dtype_repr(arr.dtype)))
-    elif is_real_dtype(arr.dtype) and not is_real_floating_dtype(arr.dtype):
-        arr = arr.astype('float64')
-
-    if axes is None:
-        axes = list(range(arr.ndim))
-    else:
-        try:
-            axes = [int(axes)]
-        except TypeError:
-            axes = list(axes)
-
-    shape = arr.shape
-    shift_list = normalized_scalar_param_list(shift, length=len(axes),
-                                              param_conv=bool)
-
-    # Make a copy of arr with correct data type if necessary, or copy values.
-    if out is None:
-        if is_real_dtype(arr.dtype) and not all(shift_list):
-            out = np.array(arr, dtype=TYPE_MAP_R2C[arr.dtype], copy=True)
-        else:
-            out = arr.copy()
-    elif out is arr:
-        pass
-    else:
-        out[:] = arr
-
-    if is_real_dtype(out.dtype) and not shift:
-        raise ValueError('cannot pre-process real input in-place without '
-                         'shift')
-
-    if sign == '-':
-        imag = -1j
-    elif sign == '+':
-        imag = 1j
-    else:
-        raise ValueError("`sign` '{}' not understood".format(sign))
-
-    def _onedim_arr(length, shift):
-        if shift:
-            # (-1)^indices
-            indices = np.arange(length)
-            arr = -2 * np.mod(indices, 2) + 1.0
-        else:
-            indices = np.arange(length)
-            arr = np.exp(-imag * np.pi * indices * (1 - 1.0 / length))
-        return arr.astype(out.dtype, copy=False)
-
-    onedim_arrs = []
-    for axis, shift in zip(axes, shift_list):
-        length = shape[axis]
-        onedim_arrs.append(_onedim_arr(length, shift))
-
-    fast_1d_tensor_mult(out, onedim_arrs, axes=axes, out=out)
-    return out
-
-
-def _interp_kernel_ft(norm_freqs, interp):
-    """Scaled FT of a one-dimensional interpolation kernel.
-
-    For normalized frequencies ``-1/2 <= xi <= 1/2``, this
-    function returns::
-
-        sinc(pi * xi)**k / sqrt(2 * pi)
-
-    where ``k=1`` for 'nearest' and ``k=2`` for 'linear' interpolation.
-
-    Parameters
-    ----------
-    norm_freqs : `numpy.ndarray`
-        Normalized frequencies between -1/2 and 1/2
-    interp : {'nearest', 'linear'}
-        Type of interpolation kernel
-
-    Returns
-    -------
-    ker_ft : `numpy.ndarray`
-        Values of the kernel FT at the given frequencies
-    """
-    # Numpy's sinc(x) is equal to the 'math' sinc(pi * x)
-    interp_ = str(interp).lower()
-    if interp_ == 'nearest':
-        return np.sinc(norm_freqs) / np.sqrt(2 * np.pi)
-    elif interp_ == 'linear':
-        return np.sinc(norm_freqs) ** 2 / np.sqrt(2 * np.pi)
-    else:
-        raise ValueError("`interp` '{}' not understood".format(interp))
-
-
-def dft_postprocess_data(arr, real_grid, recip_grid, shift, axes,
-                         interp, sign='-', op='multiply', out=None):
-    """Post-process the Fourier-space data after DFT.
-
-    This function multiplies the given data with the separable
-    function::
-
-        q(xi) = exp(+- 1j * dot(x[0], xi)) * s * phi_hat(xi_bar)
-
-    where ``x[0]`` and ``s`` are the minimum point and the stride of
-    the real-space grid, respectively, and ``phi_hat(xi_bar)`` is the FT
-    of the interpolation kernel. The sign of the exponent depends on the
-    choice of ``sign``. Note that for ``op='divide'`` the
-    multiplication with ``s * phi_hat(xi_bar)`` is replaced by a
-    division with the same array.
-
-    In discretized form on the reciprocal grid, the exponential part
-    of this function becomes an array::
-
-        q[k] = exp(+- 1j * dot(x[0], xi[k]))
-
-    and the arguments ``xi_bar`` to the interpolation kernel
-    are the normalized frequencies::
-
-        for 'shift=True'  : xi_bar[k] = -pi + pi * (2*k) / N
-        for 'shift=False' : xi_bar[k] = -pi + pi * (2*k+1) / N
-
-    See [Pre+2007]_, Section 13.9 "Computing Fourier Integrals Using
-    the FFT" for a similar approach.
-
-    Parameters
-    ----------
-    arr : `array-like`
-        Array to be pre-processed. An array with real data type is
-        converted to its complex counterpart.
-    real_grid : `RegularGrid`
-        Real space grid in the transform
-    recip_grid : `RegularGrid`
-        Reciprocal grid in the transform
-    shift : bool or `sequence` of bools
-        If ``True``, the grid is shifted by half a stride in the negative
-        direction in the corresponding axes. The sequence must have the
-        same length as ``axes``.
-    axes : int or `sequence` of ints
-        Dimensions along which to take the transform. The sequence must
-        have the same length as ``shifts``.
-    interp : string or `sequence` of strings
-        Interpolation scheme used in the real-space.
-    sign : {'-', '+'}, optional
-        Sign of the complex exponent.
-    op : {'multiply', 'divide'}
-        Operation to perform with the stride times the interpolation
-        kernel FT
-    out : `numpy.ndarray`, optional
-        Array in which the result is stored. If ``out is arr``, an
-        in-place modification is performed.
-
-    Returns
-    -------
-    out : `numpy.ndarray`
-        Result of the post-processing. If ``out`` was given, the returned
-        object is a reference to it.
-    """
-    arr = np.asarray(arr)
-    if is_real_floating_dtype(arr.dtype):
-        arr = arr.astype(TYPE_MAP_R2C[arr.dtype])
-    elif not is_complex_floating_dtype(arr.dtype):
-        raise ValueError('array data type {} is not a floating point data '
-                         'type'.format(dtype_repr(arr.dtype)))
-
-    if out is None:
-        out = arr.copy()
-    elif out is not arr:
-        out[:] = arr
-
-    if axes is None:
-        axes = list(range(arr.ndim))
-    else:
-        try:
-            axes = [int(axes)]
-        except TypeError:
-            axes = list(axes)
-
-    shift_list = normalized_scalar_param_list(shift, length=len(axes),
-                                              param_conv=bool)
-
-    if sign == '-':
-        imag = -1j
-    elif sign == '+':
-        imag = 1j
-    else:
-        raise ValueError("`sign` '{}' not understood".format(sign))
-
-    op, op_in = str(op).lower(), op
-    if op not in ('multiply', 'divide'):
-        raise ValueError("kernel `op` '{}' not understood".format(op_in))
-
-    # Make a list from interp if that's not the case already
-    try:
-        interp = [str(interp + '').lower()] * arr.ndim
-    except TypeError:
-        pass
-
-    onedim_arrs = []
-    for ax, shift, intp in zip(axes, shift_list, interp):
-        x = real_grid.min_pt[ax]
-        xi = recip_grid.coord_vectors[ax]
-
-        # First part: exponential array
-        onedim_arr = (np.exp(imag * x * xi))
-
-        # Second part: interpolation kernel
-        len_dft = recip_grid.shape[ax]
-        len_orig = real_grid.shape[ax]
-        halfcomplex = (len_dft < len_orig)
-        odd = len_orig % 2
-
-        if shift:
-            # f_k = -0.5 + k / N
-            fmin = -0.5
-            if halfcomplex:
-                if odd:
-                    fmax = - 1.0 / (2 * len_orig)
-                else:
-                    fmax = 0.0
-            else:
-                # Always -0.5 + (N-1)/N = 0.5 - 1/N
-                fmax = 0.5 - 1.0 / len_orig
-
-        else:
-            # f_k = -0.5 + 1/(2*N) + k / N
-            fmin = -0.5 + 1.0 / (2 * len_orig)
-            if halfcomplex:
-                if odd:
-                    fmax = 0.0
-                else:
-                    fmax = 1.0 / (2 * len_orig)
-            else:
-                # Always -0.5 + 1/(2*N) + (N-1)/N = 0.5 - 1/(2*N)
-                fmax = 0.5 - 1.0 / (2 * len_orig)
-
-        freqs = np.linspace(fmin, fmax, num=len_dft)
-        stride = real_grid.stride[ax]
-
-        if op == 'multiply':
-            onedim_arr *= stride * _interp_kernel_ft(freqs, intp)
-        else:
-            onedim_arr /= stride * _interp_kernel_ft(freqs, intp)
-
-        onedim_arrs.append(onedim_arr.astype(out.dtype, copy=False))
-
-    fast_1d_tensor_mult(out, onedim_arrs, axes=axes, out=out)
-    return out
-
-
-def reciprocal_space(space, axes=None, halfcomplex=False, shift=True,
-                     **kwargs):
-    """Return the range of the Fourier transform on ``space``.
-
-    Parameters
-    ----------
-    space : `DiscreteLp`
-        Real space whose reciprocal is calculated. It must be
-        uniformly discretized.
-    axes : `sequence` of ints, optional
-        Dimensions along which the Fourier transform is taken.
-        Default: all axes
-    halfcomplex : bool, optional
-        If ``True``, take only the negative frequency part along the last
-        axis for. For ``False``, use the full frequency space.
-        This option can only be used if ``space`` is a space of
-        real-valued functions.
-    shift : bool or `sequence` of bools, optional
-        If ``True``, the reciprocal grid is shifted by half a stride in
-        the negative direction. With a boolean sequence, this option
-        is applied separately to each axis.
-        If a sequence is provided, it must have the same length as
-        ``axes`` if supplied. Note that this must be set to ``True``
-        in the halved axis in half-complex transforms.
-        Default: ``True``
-    exponent : float, optional
-        Create a space with this exponent. By default, the conjugate
-        exponent ``q = p / (p - 1)`` of the exponent of ``space`` is
-        used, where ``q = inf`` for ``p = 1`` and vice versa.
-    dtype : optional
-        Complex data type of the reciprocal space. By default, the
-        complex counterpart of ``space.dtype`` is used.
-
-    Returns
-    -------
-    rspace : `DiscreteLp`
-        Reciprocal of the input ``space``. If ``halfcomplex=True``, the
-        upper end of the domain (where the half space ends) is chosen to
-        coincide with the grid node.
-    """
-    if not isinstance(space, DiscreteLp):
-        raise TypeError('`space` {!r} is not a `DiscreteLp` instance'
-                        ''.format(space))
-    if not space.is_uniform:
-        raise ValueError('`space` is not uniformly discretized')
-
-    if axes is None:
-        axes = list(range(space.ndim))
-    else:
-        try:
-            axes = [int(axes)]
-        except TypeError:
-            axes = list(axes)
-
-    if halfcomplex and space.field != RealNumbers():
-        raise ValueError('`halfcomplex` option can only be used with real '
-                         'spaces')
-
-    exponent = kwargs.pop('exponent', None)
-    if exponent is None:
-        exponent = conj_exponent(space.exponent)
-
-    dtype = kwargs.pop('dtype', None)
-    if dtype is None:
-        if is_real_dtype(space.dtype):
-            dtype = TYPE_MAP_R2C[space.dtype]
-        else:
-            dtype = space.dtype
-    else:
-        if not is_complex_floating_dtype(dtype):
-            raise ValueError('{} is not a complex data type'
-                             ''.format(dtype_repr(dtype)))
-
-    # Calculate range
-    recip_grid = reciprocal(space.grid, shift=shift, halfcomplex=halfcomplex,
-                            axes=axes)
-
-    # Make a partition with nodes on the boundary in the last transform axis
-    # if `halfcomplex == True`, otherwise a standard partition.
-    if halfcomplex:
-        max_pt = {axes[-1]: recip_grid.max_pt[axes[-1]]}
-        part = uniform_partition_fromgrid(recip_grid, max_pt=max_pt)
-    else:
-        part = uniform_partition_fromgrid(recip_grid)
-
-    recip_spc = uniform_discr_frompartition(part, exponent=exponent,
-                                            dtype=dtype, impl='numpy')
-
-    return recip_spc
 
 
 class FourierTransformBase(Operator):
