+++ odl/trafos/util/__init__.py
@@ -22,14 +22,6 @@ from __future__ import absolute_import
 
 __all__ = ()
 
-from . import util
 
-from . import backends
-from .backends.pyfftw_bindings import PYFFTW_AVAILABLE
-__all__ += (PYFFTW_AVAILABLE,)
-
-from .fourier import *
-__all__ += fourier.__all__
-
-from .wavelet import *
-__all__ += wavelet.__all__
+from .ft_utils import *
+__all__ += ft_utils.__all__
